refactor(naming): move `Capitalize` function to new `naming` package
