+++ gen/ir/json.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/ogen-go/ogen/internal/capitalize"
+	"github.com/ogen-go/ogen/internal/naming"
 	"github.com/ogen-go/ogen/jsonschema"
 )
 
@@ -124,7 +124,7 @@ func (j JSON) Format() string {
 	typePrefix := func(f string) string {
 		switch s.Type {
 		case jsonschema.String:
-			return "String" + capitalize.Capitalize(f)
+			return "String" + naming.Capitalize(f)
 		default:
 			return f
 		}
@@ -156,7 +156,7 @@ func (j JSON) Format() string {
 		if s.Type != jsonschema.String {
 			return ""
 		}
-		return "String" + capitalize.Capitalize(f)
+		return "String" + naming.Capitalize(f)
 	case "unix", "unix-seconds":
 		return typePrefix("UnixSeconds")
 	case "unix-nano":
@@ -290,7 +290,7 @@ func (j JSON) Fn() string {
 		s := j.t.Primitive.String()
 		return strings.ToUpper(s[:2]) + s[2:]
 	default:
-		return capitalize.Capitalize(j.t.Primitive.String())
+		return naming.Capitalize(j.t.Primitive.String())
 	}
 }
