+++ internal/naming/naming.go
@@ -1,0 +1,2 @@
+// Package naming contains some utilities for generating names.
+package naming
